+++ CHANGELOG.md
@@ -5,12 +5,26 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added build tasks `Generate_Wiki_Content` (that runs the cmdlet
+  `New-DscResourceWikiPage`). The build task is primarily meant to be run by
+  the project [Sampler's](https://github.com/gaelcolas/Sampler) `build.ps1`.
+  To make the task available for `Invoke-Build` in a repository that is based
+  on [Sampler](https://github.com/gaelcolas/Sampler) add this module to
+  required modules, and then in the `build.yaml` add the following.
+  ```yaml
+  ModuleBuildTasks:
+    DscResource.DocGenerator:
+      - 'Task.*'
+  ```
+
 ## [0.3.0] - 2020-02-11
 
 ### Added
 
-- Added build task `Generate_Conceptual_Help` that runs the cmdlet
-  `New-DscResourcePowerShellHelp`. The build task is primarily meant to
+- Added build tasks `Generate_Conceptual_Help` (that runs the cmdlet
+  `New-DscResourcePowerShellHelp`). The build task is primarily meant to
   be run by the project [Sampler's](https://github.com/gaelcolas/Sampler)
   `build.ps1`. To make the task available for `Invoke-Build` in a repository
   that is based on [Sampler](https://github.com/gaelcolas/Sampler) add this
@@ -20,7 +34,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     DscResource.DocGenerator:
       - 'Task.*'
   ```
-  
+
 ### Fixed
 
 - Fix the description in the README.md.